@@ -1,7 +1,8 @@
 package com.example.a18049192.oxfordtour;
 
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class Cherwell_Boathouse extends AppCompatActivity {
 
