@@ -1,7 +1,7 @@
 package com.example.a18049192.oxfordtour;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class No1ShipStreet extends AppCompatActivity {
 
