@@ -1,10 +1,10 @@
 package com.example.a18049192.oxfordtour;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class PlacesPageMenu extends AppCompatActivity {
 
